Adding dockerfile to create the image to run the container. Some improvements in the logic of extracting CV information and visual improvements to the web page with streamlit
