--- app/DockerFile
@@ -1,27 +1,0 @@
-# Usa una imagen oficial de Python
-FROM python:3.11-slim
-
-# Evita problemas con entrada/salida en algunos entornos
-ENV PYTHONUNBUFFERED=1
-
-# Instala dependencias del sistema
-RUN apt-get update && apt-get install -y \
-    build-essential \
-    libpoppler-cpp-dev \
-    curl \
-    && rm -rf /var/lib/apt/lists/*
-
-# Crea una carpeta en el contenedor
-WORKDIR /app
-
-# Copia los archivos del proyecto al contenedor
-COPY . .
-
-# Instala las dependencias de Python
-RUN pip install --no-cache-dir -r requirements.txt
-
-# Expone el puerto que usa Streamlit
-EXPOSE 8501
-
-# Comando por defecto: correr Streamlit
-CMD ["streamlit", "run", "app.py", "--server.port=8501", "--server.address=0.0.0.0"]
